Create string_concat() for Task 1
Use malloc to allocate space matching the size of input strings.
Place if conditional right after to ensure that no operations
dealing with memory are made if no memory was actually allocated.
Create for loop to loop through ea input string and assign each char
in strings to allocated memory.