Modify algo for insertion_sort.
Initialize the index @ind at 0 instead of size - 1, and increment
ind each round instead of decrementing. This causes sorting to be-
gin with just one index of the array and increase the indices to
be sorted with each round.

Remove var @r since @ind can be used for the same purpose (repre-
senting how many rounds of sorting to make with the conditional
ind < size).

Test; code works.