Write out verbal pseudocode for quick_sort.
Proposed algorithm entails partitioning of array, calculating size
of left-most subarray, and repeating steps so long as subarray
size is greater than 2; in a left-to-right fashion. This seems to
require storing the indices that have served as pivots into an new
array, so that the next subarray to partition down into two is
known.